pyproject.toml
[project]
name = "csfloat-monitor"
version = "0.1.0"
description = "CSFloat listing monitor with persistent state tracking and Telegram alerts"
readme = "README.md"
requires-python = ">=3.13.12"
dependencies = [
    "httpx>=0.28.1",
    "peewee>=4.0.3",
    "peewee-db-evolve>=3.7.6",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.2.2",
]
